requirements.txt
fastapi
starlette-exporter==0.12.0
uvicorn[standard]
pandas
aiofiles